+++ plugins/legacy/History.md
@@ -1,5 +1,11 @@
 # Changelog for expresser-legacy
 
+3.6.1
+=====
+* Fixes for using Connect Assets with coffeescript.
+* Removed debugging console calls.
+* Removed process.exit() log.
+
 3.6.0
 =====
 * Initial adapter release.